feat(source): order units by recency, subagents last
JsonlTree::units() now always sorts by mtime desc (not only under --limit), then
stable-sorts subagents last, so the queue is main-first, recency-desc within each
group. --limit still keeps the recent N (truncate before the subagent sort). Full-
index store output is unchanged — chunk-id dedup makes it order-independent — so this
only sets the consumption order the seed and budgeted backfill will use.

Unit gains is_subagent, flagged via jsonl::is_subagent. That predicate (Claude names
sub-agents `agent-<hash>.jsonl`) moves from curate to jsonl beside session_id_of, so
indexing and curation share one copy instead of two that can drift.
